refactor(vec): move Vec logic into BaseVec
This change factors StableVec functionality into an internal BaseVec
type parameterized over the stable structure magic number.

The motivation is to re-use the BaseVec code in more specialized
array-like data structures, such as binary heaps and certified
sequences.
